Yamaha YSFC: fix the envelope time quantization, the doubled tuning and duplicated performances (#262)
* Yamaha YSFC: fix the envelope time quantization, the doubled tuning and duplicated performances

The lookup from seconds into the envelope time table compared against
the truncated integer value of the table entries (which range from 0.2
to 90 seconds), so every sub-second time landed on the first entry
whose neighbour reaches 1 second - about 0.9 seconds - in all written
AEG, FEG and PEG times. The tuning of a zone was written to both the
keybank and the element although the device and the reader sum the
two, which doubled the detune; the element tuning is now left neutral.
When reading performances, the performance source was added to the
result list inside the part loop instead of after it, so a multi-part
performance was output once per part.

* Move the changelog entry to the collected changelog PR #283

The open fix PRs all inserted their changelog entries near the same
lines, so every merged PR turned the remaining ones into conflicts.
The entries of the whole batch are collected in PR #283 and quoted in
each PR's description; the fix branches no longer touch the changelog,
so they merge cleanly in any order.

+++ src/main/java/de/mossgrabers/convertwithmoss/format/yamaha/ysfc/file/YamahaYsfcPartElement.java
@@ -2262,7 +2262,7 @@ public class YamahaYsfcPartElement
             return 0;
 
         for (int i = 0; i < 127; i++)
-            if (seconds >= ENVELOPE_TIMES.get (Integer.valueOf (i)).intValue () && seconds < ENVELOPE_TIMES.get (Integer.valueOf (i + 1)).intValue ())
+            if (seconds >= ENVELOPE_TIMES.get (Integer.valueOf (i)).doubleValue () && seconds < ENVELOPE_TIMES.get (Integer.valueOf (i + 1)).doubleValue ())
                 return i;
 
         return 127;
